+++ __tests__/functions/validateFunction.test.ts
@@ -2,7 +2,7 @@ import test, { ExecutionContext } from 'ava';
 import path from 'path';
 import { Validator } from 'jsonschema';
 import { functionDefinition } from '../../src/functions/functionDefinitions';
-import { validateFunction } from '../../src/functions/validations';
+import { validateSchema } from '../../src/functions/validations';
 
 type Context = ExecutionContext<unknown>;
 
@@ -10,31 +10,29 @@ const schema = {
   $id: '/schema/actions/function.json',
   title: 'Function',
   properties: {
-    name: {
+    label: {
       type: 'string',
     },
     icon: {
       enum: ['CreateIcon', 'DeleteIcon'],
     },
   },
-  required: ['name'],
+  required: ['label'],
 };
 
 const validator = new Validator();
 validator.addSchema(schema, schema.$id);
 
-test('load in entire schema for validator', async (t: Context): Promise<
-  void
-> => {
+test('load in entire schema for validator', async (t: Context): Promise<void> => {
   const definition = {
     path: '/path/to/schema/actions/function.json',
     schema: {
-      name: 'create',
+      label: 'Create',
       icon: 'CreateIcon',
     },
   };
 
-  const { status, errors } = await validateFunction(definition, validator);
+  const { status, errors } = await validateSchema(definition, validator);
 
   t.is(status, 'ok');
   t.is(errors.length, 0);
@@ -44,11 +42,15 @@ test('validate templates', async (t: Context): Promise<void> => {
   const functionPath = path.join(
     process.cwd(),
     'assets/app-functions/templates',
-    'functions/say-hello',
+    'functions/say-hello/1.0',
   );
 
-  const functionJson = functionDefinition(functionPath);
-  const { status } = await validateFunction(functionJson, validator);
+  const functionJson = functionDefinition(
+    functionPath,
+    path.join(process.cwd(), 'functions'),
+  );
+
+  const { status } = await validateSchema(functionJson, validator);
 
   t.is(status, 'ok');
 });
@@ -57,7 +59,7 @@ test('invalidate empty schemas', async (t: Context): Promise<void> => {
   const {
     status,
     errors: [{ message }],
-  } = await validateFunction(
+  } = await validateSchema(
     {
       path: '/path/to/schema/actions/function.json',
       schema: {},
@@ -66,16 +68,14 @@ test('invalidate empty schemas', async (t: Context): Promise<void> => {
   );
 
   t.is(status, 'error');
-  t.is(message, 'requires property "name"');
+  t.is(message, 'requires property "label"');
 });
 
-test('invalidate schemas that do not have valid values for properties', async (t: Context): Promise<
-  void
-> => {
+test('invalidate schemas that do not have valid values for properties', async (t: Context): Promise<void> => {
   const definition = {
     path: '/path/to/schema/actions/function.json',
     schema: {
-      name: 'create',
+      label: 'Create',
       icon: 'RandomIcon',
     },
   };
@@ -83,7 +83,7 @@ test('invalidate schemas that do not have valid values for properties', async (t
   const {
     status,
     errors: [{ message }],
-  } = await validateFunction(definition, validator);
+  } = await validateSchema(definition, validator);
 
   t.is(status, 'error');
   t.is(message, 'is not one of enum values: CreateIcon,DeleteIcon');