+++ src/bb-functions-validate.ts
@@ -1,5 +1,6 @@
 /* npm dependencies */
 
+import chalk from 'chalk';
 import path from 'path';
 import program from 'commander';
 
@@ -9,6 +10,9 @@ import {
   FunctionValidator,
   logValidationResult,
 } from './functions/validations';
+
+import { functionDefinitions } from './functions/functionDefinitions';
+
 import Config from './functions/config';
 
 /* process arguments */
@@ -32,11 +36,17 @@ void (async (): Promise<void> => {
   const validator = new FunctionValidator(config, baseFunctionsPath);
   await validator.initSchema();
 
-  if (inputFunctionName) {
-    const result = validator.validateFunction(inputFunctionName);
-    logValidationResult(result);
-  } else {
-    const results = await validator.validateFunctions();
-    results.forEach((result) => logValidationResult(result));
+  const results = await validator.validateFunctions(inputFunctionName);
+  results.forEach(logValidationResult);
+
+  const allFunctions = functionDefinitions(baseFunctionsPath, true);
+  const versionedFunctions = functionDefinitions(baseFunctionsPath);
+
+  if (allFunctions.length !== versionedFunctions.length) {
+    console.log(
+      `Maybe auto-version your functions without a version number using ${chalk.cyan(
+        'bb functions autoversion',
+      )}?`,
+    );
   }
 })();