+++ src/bb-functions.ts
@@ -16,6 +16,8 @@ const availableCommands: CommandFunctions[] = [
   'build',
   'publish',
   'validate',
+  'autoversion',
+  'bump',
   'test',
 ];
 
@@ -24,11 +26,16 @@ const availableCommands: CommandFunctions[] = [
 program
   .usage(`<${availableCommands.join('|')}>`)
   .name('bb functions')
+  .command(
+    'autoversion',
+    'auto-version functions within current working directory',
+  )
+  .command('build', 'build functions bundle file of current working directory')
+  .command('bump', 'increase the major/minor version of a specific function')
   .command('init [identifier]', 'initialize functions project')
   .command('login', 'login using the same credentials as the IDE')
   .command('logout', 'remove all tokens used to authenticate with the APIs')
   .command('new [function-name]', 'Initialize a new function')
-  .command('build', 'build functions bundle file of current working directory')
   .command('publish', 'publish functions of current working directory')
   .command('validate', 'validate functions of current working directory')
   .command(