+++ src/types.ts
@@ -25,6 +25,8 @@ export type CommandFunctions =
   | 'build'
   | 'publish'
   | 'validate'
+  | 'autoversion'
+  | 'bump'
   | 'test';
 
 export type CommandInteractions = 'generate';
@@ -49,9 +51,15 @@ export interface Component {
   type: string;
 }
 
+export type PrefabReference = PrefabComponent | PrefabPartial;
+
+export type PrefabPartial = {
+  type: 'PARTIAL';
+  partialId: string;
+};
 export interface PrefabComponent {
+  type?: 'COMPONENT';
   actions?: PrefabAction[];
-  descendants: PrefabComponent[];
   name: string;
   style?: {
     name?: string;
@@ -83,6 +91,7 @@ export interface PrefabComponent {
       textTransform?: string;
     };
   };
+  descendants: PrefabReference[];
   options: PrefabComponentOption[];
   ref?: {
     id: string;
@@ -160,7 +169,7 @@ export interface Prefab {
   keywords?: string[];
   icon: Icon;
   interactions?: PrefabInteraction[];
-  structure: PrefabComponent[];
+  structure: PrefabReference[];
   variables?: PrefabVariable[];
   type?: string;
   description?: string;