+++ src/validations/prefab/component.ts
@@ -6,14 +6,15 @@ import Joi from 'joi';
 
 import {
   Prefab,
-  PrefabComponent,
   ComponentStyleMap,
   Component,
+  PrefabReference,
 } from '../../types';
 import { findDuplicates } from '../../utils/validation';
 import { optionSchema } from './componentOption';
 
 type StyleValidator = Record<Component['styleType'], Joi.ObjectSchema>;
+type PrefabTypes = 'partial' | 'page' | undefined;
 
 const shadows = [
   'none',
@@ -85,9 +86,17 @@ const styleValidator: StyleValidator = {
   }),
 };
 
+const partialSchema = (): Joi.ObjectSchema => {
+  return Joi.object({
+    type: Joi.string().valid('PARTIAL').required(),
+    partialId: Joi.string().allow('').required(),
+  });
+};
+
 const componentSchema = (
   componentStyleMap?: ComponentStyleMap,
   styleType?: keyof StyleValidator,
+  prefabType?: PrefabTypes,
 ): Joi.ObjectSchema => {
   const canValidateStyle = styleType && styleValidator[styleType];
 
@@ -101,8 +110,9 @@ const componentSchema = (
       id: Joi.string().required(),
     }),
     options: Joi.array().items(optionSchema).required(),
+    type: Joi.string().valid('COMPONENT').default('COMPONENT'),
     descendants: Joi.array()
-      .items(Joi.custom(validateComponent(componentStyleMap)))
+      .items(Joi.custom(validateComponent(componentStyleMap, prefabType)))
       .required(),
 
     // lifecycle hooks
@@ -159,27 +169,46 @@ const componentSchema = (
 };
 
 export const validateComponent =
-  (componentStyleMap?: ComponentStyleMap) =>
-  (component: PrefabComponent): Prefab | unknown => {
-    const { name, options } = component;
+  (componentStyleMap?: ComponentStyleMap, prefabType?: PrefabTypes) =>
+  (component: PrefabReference): Prefab | unknown => {
+    if (component.type === 'PARTIAL') {
+      if (prefabType === 'partial') {
+        throw new Error(
+          chalk.red(`\n Partials are not supported in partial Prefabs\n`),
+        );
+      }
+      const { type } = component;
+      const { error } = partialSchema().validate(component);
 
-    const styleType: Component['styleType'] | undefined =
-      componentStyleMap &&
-      componentStyleMap[name] &&
-      componentStyleMap[name].styleType;
-    const { error } = componentSchema(
-      componentStyleMap,
-      styleType as keyof StyleValidator,
-    ).validate(component);
+      if (typeof error !== 'undefined') {
+        const { message } = error;
 
-    findDuplicates(options, 'option key', 'key');
+        throw new Error(
+          chalk.red(`\nBuild error in component ${type}: ${message}\n`),
+        );
+      }
+    } else {
+      const { name, options } = component;
 
-    if (typeof error !== 'undefined') {
-      const { message } = error;
+      const styleType: Component['styleType'] | undefined =
+        componentStyleMap &&
+        componentStyleMap[name] &&
+        componentStyleMap[name].styleType;
+      const { error } = componentSchema(
+        componentStyleMap,
+        styleType as keyof StyleValidator,
+        prefabType,
+      ).validate(component);
 
-      throw new Error(
-        chalk.red(`\nBuild error in component ${name}: ${message}\n`),
-      );
+      findDuplicates(options, 'option key', 'key');
+
+      if (typeof error !== 'undefined') {
+        const { message } = error;
+
+        throw new Error(
+          chalk.red(`\nBuild error in component ${name}: ${message}\n`),
+        );
+      }
     }
 
     return component;